Whitelist IP of Ansible controller via UFW
If "limit" rule is applied and controller exceeds the connection
threshold, it becomes possible to lock oneself out of the target host.

This commit adds `base_operator_ip_whitelist` variable which, if set to
True, will create an "allow" rule to allow all traffic from the current
public IP of the controller.

Closes https://github.com/savchenko/debian/issues/18